requirements.txt
python-dotenv==1.1.1
slack-bolt==1.23.0
fastapi==0.115.6
uvicorn==0.32.1
requests==2.32.3
